baseline graphs
baselines graphs uses directlingam in cuda. Also graph comparison was tried but the results are bad. There must be something else related that we should be able to do to get good results. the proccess is simple: get the baseline graph only with benign traffic and the other graph only using malicious traffic. then send batches of data to create a new graph and compare to the baseline ones and then decide if there is malicious traffic on that batch